+(*)Use tv%SpV in MOM_sponge code
  Use tv%SpV convert thicknesses to vertical distances in apply_sponge when it
is allocated to replace multiplication by GV%H_to_Z, thereby eliminating the
dependence on GV%RHo_0 in this modue in non-Boussinesq mode.  The new internal
array dz_to_h is used to reduce the code duplication as a result of these
changes.  All answers in Boussinesq test cases are bitwise identical, but
answers change in fully non-Boussinesq mode.  In semi-Boussinesq mode answers
are mathematically equivalent but change at roundoff unless RHO_0 is an integer
power of 2.