Autoconf: Fix Python test and allow configuration
The AC_PATH_PROGS macros used in Python testing were incorrectly using
AC_MSG_ERROR in places where a missing value for PYTHON should be if the
executable was not found.

It also did not permit for a configurable PYTHON variable, since the
autodetect was always run, even if PYTHON were set.

This has been updated so that Python autodetection only runs if
PYTHON is unset.  It also correctly reports a failed configuration if
PYTHON is not found.

(It does not, however, test of PYTHON is actually a Python interpreter,
but we can deal with that at a later date.)