*Revise calc_isoneutral_slopes when non-Boussinesq
  Revise calc_isoneutral_slopes to eliminate any dependence on the Boussinesq
reference density when in non-Boussinesq mode.  This includes the addition of
two new arrays to hold the rescaled product of the gravitational acceleration
and the specific volume interpolated to the interfaces at velocity points.
The answers change in non-Boussinesq mode when one of several parameterizations
that use the isoneutral slopes are in use, but are bitwise identical in
Boussinesq or semi-Boussinesq mode.