+Use RESTORE_FLUX_RHO in dumbbell & SCM_CVMix_tests
  Use RESTORE_FLUX_RHO in SCM_CVMix_tests and dumbbell_surface_forcing to
specify the density that are used to convert the piston velocities into
restoring heat or salt fluxes.  As with other analogous changes, the default is
set to RHO_0 to reproduce previous answers.  Also set forces%tau_mag in
SCM_CVMix_tests_wind_forcing if it is associated.  There is a new rho_restore
element in the control structures for the SCM_CVMix_tests two module, while the
units of an element in the dumbbell_surface_forcing module are changed. By
default, all existing answers are bitwise identical, but there are new entries
in some MOM_parameter_doc files.