*Revise BFB_set_coord and BFB_buoyancy_forcing
  Revised BFB_set_coord and BFB_buoyancy_forcing to be consistent with other
instances of a linear equation of state, and to set g_prime in non-Boussinesq
mode similarly to how it is set in other places.   Also use RESTORE_FLUX_RHO in
place of RHO_0 when to specify the densities that are used to convert the piston
velocities into restoring heat or salt fluxes with BFB_buoyancy_forcing, like in
other places in the MOM6 code.  This change will change answers by default when
BFB_set_coord is used, but the old Boussinesq answers can be recovered by
setting the reference salinity S_REF to 38.75 ppt or by setting RHO_T0_S0 to
1003.0 kg m-3.