*Correct nuopc_cap tau_mag bug
  Correct a recently added bug in the expression for tau_mag in the nuopc_cap
version of convert_IOB_to_forces, where CS%gust(i,j) was used in place of
CS%gust_const, even though the 2-d array was not being set.  This commit changes
answers in some recent versions of the code back to what they had been
previously, and it addresses concerns that had been raised with the first
version of gfdl-candidate-2023-07-03 and its PR to the main version of MOM6.