+*Revise non-Boussinesq offline tracer diffusivity
  Rescale the units diapycnal diffusivity used in the offline tracer mode to
work in [H Z T-1 ~> m2 s-1 or kg m-1 s-1], making it consistent with what has
already been done for other diapycnal diffusivities.  This change involves
changes to the units of arguments to update_offline_from_files and
ALE_offline_inputs and to two elements of offline_transport_CS.   It also
includes a new thermo_vary_type argument to offline_diabatic_ale, as well as a
call to thickness_to_dz in that same routine to set the values of the newly
added 3-d array of the vertical distances across layers and a change in name and
units of the inverse thickness variable used to find the fluxes.  All answers
are bitwise identical in Boussinesq mode, but in offline tracer calculations in
non-Boussinesq mode they will change to become independent of the Boussinesq
reference density.