Make HBD work with 3D diffusivities
Following up on the previous commit, where a vertical dimension
was added to the tracer diffusivities, this commit modifies the
HBD module to work with this change. To do so, parameter khtr_u
(diffusivity times the time step) is calculated at cell centers
and then remapped onto the HBD vertical grid. All unit tests
in this module were updated to conform with this change.

This commit also makes the default value of HBD_DEBUG equal to
the value set for DEBUG.